Another refactor, remove linux specific code as test
